core/src/test/java/nl/matsgemmeke/battlegrounds/item/equipment/DefaultEquipmentTest.java
package nl.matsgemmeke.battlegrounds.item.equipment;

import nl.matsgemmeke.battlegrounds.item.ItemTemplate;
import nl.matsgemmeke.battlegrounds.item.controls.Action;
import nl.matsgemmeke.battlegrounds.item.controls.ItemFunction;
import nl.matsgemmeke.battlegrounds.item.deployment.Deployable;
import org.bukkit.Material;
import org.bukkit.inventory.ItemStack;
import org.junit.Test;

import static org.junit.Assert.*;
import static org.mockito.Mockito.*;

public class DefaultEquipmentTest {

    @Test
    public void shouldDoNothingIfHolderIsNullWhenDeploying() {
        Deployable object = mock(Deployable.class);

        DefaultEquipment equipment = new DefaultEquipment();
        equipment.onDeploy(object);

        assertTrue(equipment.getDeployedObjects().isEmpty());
    }

    @Test
    public void shouldSetHolderHeldItemToActivatorItemStackWhenDeploying() {
        Deployable object = mock(Deployable.class);
        EquipmentHolder holder = mock(EquipmentHolder.class);
        ItemStack activatorItemStack = new ItemStack(Material.SHEARS);

        ItemTemplate activatorItemTemplate = mock(ItemTemplate.class);
        when(activatorItemTemplate.createItemStack(any())).thenReturn(activatorItemStack);

        DefaultEquipment equipment = new DefaultEquipment();
        equipment.setActivatorItemTemplate(activatorItemTemplate);
        equipment.setHolder(holder);
        equipment.onDeploy(object);

        assertEquals(1, equipment.getDeployedObjects().size());
        assertEquals(object, equipment.getDeployedObjects().get(0));

        verify(holder).setHeldItem(activatorItemStack);
    }

    @Test
    @SuppressWarnings("unchecked")
    public void shouldPerformFunctionWhenLeftClicked() {
        EquipmentHolder holder = mock(EquipmentHolder.class);

        ItemFunction<EquipmentHolder> function = (ItemFunction<EquipmentHolder>) mock(ItemFunction.class);
        when(function.isAvailable()).thenReturn(true);

        DefaultEquipment equipment = new DefaultEquipment();
        equipment.getControls().addControl(Action.LEFT_CLICK, function);
        equipment.setHolder(holder);
        equipment.onLeftClick();

        verify(function).perform(holder);
    }

    @Test
    @SuppressWarnings("unchecked")
    public void shouldPerformFunctionWhenRightClicked() {
        EquipmentHolder holder = mock(EquipmentHolder.class);

        ItemFunction<EquipmentHolder> function = (ItemFunction<EquipmentHolder>) mock(ItemFunction.class);
        when(function.isAvailable()).thenReturn(true);

        DefaultEquipment equipment = new DefaultEquipment();
        equipment.getControls().addControl(Action.RIGHT_CLICK, function);
        equipment.setHolder(holder);
        equipment.onRightClick();

        verify(function).perform(holder);
    }

    @Test
    public void doesNotUpdateIfItemTemplateIsNull() {
        DefaultEquipment equipment = new DefaultEquipment();
        equipment.setItemTemplate(null);
        boolean updated = equipment.update();

        assertFalse(updated);
    }

    @Test
    public void createNewItemStackFromTemplateWhenUpdatingAndSetHeldItemOfHolder() {
        EquipmentHolder holder = mock(EquipmentHolder.class);
        ItemStack itemStack = new ItemStack(Material.SHEARS);

        ItemTemplate itemTemplate = mock(ItemTemplate.class);
        when(itemTemplate.createItemStack(any())).thenReturn(itemStack);

        DefaultEquipment equipment = new DefaultEquipment();
        equipment.setHolder(holder);
        equipment.setItemTemplate(itemTemplate);
        equipment.setName("test");
        equipment.setItemStack(itemStack);
        boolean updated = equipment.update();

        assertTrue(updated);

        verify(holder).setHeldItem(itemStack);
    }
}
